.pub-cache/hosted/pub.dartlang.org/html-0.15.0/test/data/tree-construction/scripted/webkit01.dat

#data
1<script>document.write("2")</script>3
#errors
#document
| <html>
|   <head>
|   <body>
|     "1"
|     <script>
|       "document.write("2")"
|     "23"

#data
1<script>document.write("<script>document.write('2')</scr"+ "ipt><script>document.write('3')</scr" + "ipt>")</script>4
#errors
#document
| <html>
|   <head>
|   <body>
|     "1"
|     <script>
|       "document.write("<script>document.write('2')</scr"+ "ipt><script>document.write('3')</scr" + "ipt>")"
|     <script>
|       "document.write('2')"
|     "2"
|     <script>
|       "document.write('3')"
|     "34"
